# Speed up batch generation for tall data frames