@ali-tny 2020-09-11
It looks like view vs copy in pandas is a bit of a quagmire - the documentation doesn't seem to guarantee even that[ iloc returns a view:](https://pandas.pydata.org/pandas-docs/stable/user_guide/indexing.html#indexing-view-versus-copy) 

> Outside of simple cases, it’s very hard to predict whether it will return a view or a copy (it depends on the memory layout of the array, about which pandas makes no guarantees)

playing around it looks like it depends on whether the dataframe has multiple types in it? which is presumably fairly common.

so - can you confirm this doesn't degrade performance for wide DFs? or perhaps there's a way we could do this in a single indexing operation.
Alternatively if neither of the above are true then I'd probably be happy to change it anyway since it was probably a pretty arbitrary decision to do the column selection before in the first place