@ig248 2020-09-11
This is actually as convoluted as you say :D I am guessing that one of the issues I encountered is that `df[df.columns]` can be very slow. I will try  to add a few benchmarks to the PR.