@ali-tny 2020-09-11
also the usecase that rows >>>>> columns is probably vastly more common (given that the batching is being done by row) so the wide DF case is unlikely to matter too much